feat: cli plugin
