+++ connector/internal/types.go
@@ -24,46 +24,6 @@ var (
 
 var defaultRetryHTTPStatus = []int64{429, 500, 502, 503}
 
-const (
-	RESTOptionsArgumentName          string = "restOptions"
-	RESTSingleOptionsObjectName      string = "RestSingleOptions"
-	RESTDistributedOptionsObjectName string = "RestDistributedOptions"
-	RESTServerIDScalarName           string = "RestServerId"
-	DistributedErrorObjectName       string = "DistributedError"
-)
-
-// SingleObjectType represents the object type of REST execution options for single server
-var SingleObjectType rest.ObjectType = rest.ObjectType{
-	Description: utils.ToPtr("Execution options for REST requests to a single server"),
-	Fields: map[string]rest.ObjectField{
-		"servers": {
-			ObjectField: schema.ObjectField{
-				Description: utils.ToPtr("Specify remote servers to receive the request. If there are many server IDs the server is selected randomly"),
-				Type:        schema.NewNullableType(schema.NewArrayType(schema.NewNamedType(RESTServerIDScalarName))).Encode(),
-			},
-		},
-	},
-}
-
-// DistributedObjectType represents the object type of REST execution options for distributed servers
-var DistributedObjectType rest.ObjectType = rest.ObjectType{
-	Description: utils.ToPtr("Distributed execution options for REST requests to multiple servers"),
-	Fields: map[string]rest.ObjectField{
-		"servers": {
-			ObjectField: schema.ObjectField{
-				Description: utils.ToPtr("Specify remote servers to receive the request"),
-				Type:        schema.NewNullableType(schema.NewArrayType(schema.NewNamedType(RESTServerIDScalarName))).Encode(),
-			},
-		},
-		"parallel": {
-			ObjectField: schema.ObjectField{
-				Description: utils.ToPtr("Execute requests to remote servers in parallel"),
-				Type:        schema.NewNullableNamedType(string(rest.ScalarBoolean)).Encode(),
-			},
-		},
-	},
-}
-
 // RESTOptions represent execution options for REST requests
 type RESTOptions struct {
 	Servers  []string `json:"servers"  yaml:"serverIds"`
