+++ connector/types.go
@@ -5,14 +5,12 @@ import (
 	"net/http"
 
 	"github.com/hasura/ndc-rest/connector/internal"
-	rest "github.com/hasura/ndc-rest/ndc-rest-schema/schema"
 )
 
 var (
 	errInvalidSchema      = errors.New("failed to validate NDC REST schema")
 	errBuildSchemaFailed  = errors.New("failed to build NDC REST schema")
 	errHTTPMethodRequired = errors.New("the HTTP method is required")
-	errFilePathRequired   = errors.New("file path is empty")
 )
 
 // State is the global state which is shared for every connector request.
@@ -36,10 +34,4 @@ func WithClient(client internal.Doer) Option {
 	return func(opts *options) {
 		opts.client = client
 	}
-}
-
-// NDCRestSchemaWithName wraps NDCRestSchema with identity name
-type NDCRestSchemaWithName struct {
-	name   string
-	schema *rest.NDCRestSchema
 }
