+++ ndc-rest-schema/schema/constant.go
@@ -1,0 +1,9 @@
+package schema
+
+const (
+	RESTOptionsArgumentName          string = "restOptions"
+	RESTSingleOptionsObjectName      string = "RestSingleOptions"
+	RESTDistributedOptionsObjectName string = "RestDistributedOptions"
+	RESTServerIDScalarName           string = "RestServerId"
+	DistributedErrorObjectName       string = "DistributedError"
+)
